Computes one or several polynomials to represent a given neural network using the NN2Poly algorithm.

Performs the full NN2Poly algorithm that obtains polynomial coefficients for a model that performs closely as a given already trained neural network using its weights and a Taylor approximation of its activation functions.

Usage

nn2poly_algorithm(
  weights_list,
  af_string_list,
  q_taylor_vector,
  all_partitions = NULL,
  store_coeffs = FALSE,
  forced_max_Q = NULL
)

Arguments

weights_list

list of length L ( number of hidden layers + 1) containing the weights matrix for each layer. The expected shape of such matrices at any layer L is of the form $(h_(l-1) + 1)*(h_l)$, that is, the number of rows is the number of neurons in the previous layer plus one (as the bias vector is added in the first row), and the number of columns is the number of neurons in the current layer L. Therefore, each column corresponds to the weight vector affecting each neuron in that layer, from 0 (the bias) in row 1, to neuron h_l in row h_l +1.

af_string_list

list of length L containing character strings with the names of the activation function used at each layer.

q_taylor_vector

vector of length L containing the degree (numeric) up to which Taylor expansion should be performed at each layer.

all_partitions

Optional argument containing the needed multipartitions as list of lists of lists. If NULL, the function computes it first. This step can be computationally expensive and it is encouraged that the multipartitions are stored and reused when possible.

store_coeffs

Boolean that determines if all polynomials computed in the internal layers have to be stored and given in the output (TRUE), or if only the last layer is needed (FALSE). Default is FALSE.

forced_max_Q

Optional argument: integer that determines the maximum order that we will force in the final polynomial, discarding terms of higher order that would naturally arise using all the orders in q_taylor_vector.

Value

If store_coeffs = FALSE (default case), it returns a list with an item named labels that is a list of integer vectors with each the variables index associated to each polynomial term, and a item named values

which contains a matrix where each row are the coefficients of the polynomial associated with an output neuron.

If store_coeffs = TRUE, it returns a list of length L that for each layer contains an item as explained before. The polynomials obtained at the hidden layers are not needed to represent the NN but can be used to explore how the method works.