Evaluates one or several polynomials over one or more data points in the desired variables.

Evaluates one or several polynomials over one or more data points in the desired variables.

Usage

eval_poly(x, poly)

Arguments

x

Input data as matrix, vector or dataframe. The number of columns should be the number of variables in the polynomial (the dimension p). Response variable to be predicted should not be included.

poly

List containing 2 items: labels and values.

labels List of integer vectors with same length (or number of rows) as values, where each integer vector denotes the combination of variables associated to the coefficient value stored at the same position in values. Note that the variables are numbered from 1 to p.

values Matrix (or also a vector if single polynomial), where each row represents a polynomial, with same number of columns as the length of labels, containing at each column the value of the coefficient given by the equivalent label in that same position.

Example: If labels contains the integer vector c(1,1,3) at position 5, then the value stored in values at position 5 is the coefficient associated with the term x_1^2*x_3.

Value

Vector containing the evaluation of a single polynomial or matrix containing the evaluation of the polynomials. Each row corresponds to each polynomial used and each column to each observation, meaning that each row vector corresponds to the results of evaluating all the given data for each polynomial.

Examples

# Create the polynomial 1 + (-1)·x_1 + 1·x_2 + 0.5·(x_1)^2 as a list
poly <- list()
poly$values <- c(1,-1,1,0.5)
poly$labels <- list(c(0),c(1),c(2),c(1,1))
# Create two observations, (x_1,x_2) = (1,2) and (x_1,x_2) = (3,1)
x <- rbind(c(1,2), c(3,1))
# Evaluate the polynomial on both observations
eval_poly(x,poly)
#> [1] 2.5 3.5