Predict method for nn2poly objects.

Predicted values obtained with a nn2poly object on given data.

Usage

# S3 method for nn2poly
predict(object, newdata, layers = NULL, ...)

Arguments

object

Object of class inheriting from 'nn2poly'.

newdata

Input data as matrix, vector or dataframe. Number of columns (or elements in vector) should be the number of variables in the polynomial (dimension p). Response variable to be predicted should not be included.

layers

Vector containing the chosen layers from object to be evaluated. If set to NULL, all layers are computed. Default is set to NULL.

...

Further arguments passed to or from other methods.

Value

Returns a matrix or list of matrices with the evaluation of each polynomial at each layer as given by the provided object of class nn2poly.

If object contains the polynomials of the last layer, as given by nn2poly(object, keep_layers = FALSE), then the output is a matrix with the evaluation of each data point on each polynomial. In this matrix, each column represents the evaluation of a polynomial and each column corresponds to each point in the new data to be evaluated.

If object contains all the internal polynomials also, as given by nn2poly(object, keep_layers = TRUE), then the output is a list of layers (represented by layer_i), where each one is another list with input and output elements, where each one contains a matrix with the evaluation of the "input" or "output" polynomial at the given layer, as explained in the case without internal polynomials.

Details

Internally uses eval_poly() to obtain the predictions. However, this only works with a objects of class nn2poly while eval_poly() can be used with a manually created polynomial in list form.

When object contains all the internal polynomials also, as given by nn2poly(object, keep_layers = TRUE), it is important to note that there are two polynomial items per layer (input/output). These polynomial items will also contain several polynomials of the same structure, one per neuron in the layer, stored as matrix rows in $values. Please see the NN2Poly original paper for more details.

Note also that "linear" layers will contain the same input and output results as Taylor expansion is not used and thus the polynomials are also the same. Because of this, in the situation of evaluating multiple layers we provide the final layer with "input" and "output" even if they are the same, for consistency.

See also

Examples

# Build a NN structure with random weights, with 2 (+ bias) inputs,
# 4 (+bias) neurons in the first hidden layer with "tanh" activation
# function, 4 (+bias) neurons in the second hidden layer with "softplus",
# and 1 "linear" output unit

weights_layer_1 <- matrix(rnorm(12), nrow = 3, ncol = 4)
weights_layer_2 <- matrix(rnorm(20), nrow = 5, ncol = 4)
weights_layer_3 <- matrix(rnorm(5), nrow = 5, ncol = 1)

# Set it as a list with activation functions as names
nn_object = list("tanh" = weights_layer_1,
                 "softplus" = weights_layer_2,
                 "linear" = weights_layer_3)

# Obtain the polynomial representation (order = 3) of that neural network
final_poly <- nn2poly(nn_object, max_order = 3)

# Define some new data, it can be vector, matrix or dataframe
newdata <- matrix(rnorm(10), ncol = 2, nrow = 5)

# Predict using the obtained polynomial
predict(object = final_poly, newdata = newdata)
#> [1] 154497677655    200445286  10259082512    648850668  -6622114287

# Change the last layer to have 3 outputs (as in a multiclass classification)
# problem
weights_layer_4 <- matrix(rnorm(20), nrow = 5, ncol = 4)

# Set it as a list with activation functions as names
nn_object = list("tanh" = weights_layer_1,
                 "softplus" = weights_layer_2,
                 "linear" = weights_layer_4)

# Obtain the polynomial representation of that neural network
# Polynomial representation of each hidden neuron is given by
final_poly <- nn2poly(nn_object, max_order = 3, keep_layers = TRUE)

# Define some new data, it can be vector, matrix or dataframe
newdata <- matrix(rnorm(10), ncol = 2, nrow = 5)

# Predict using the obtained polynomials (for all layers)
predict(object = final_poly, newdata = newdata)
#> $layer_1
#> $layer_1$input
#>            [,1]     [,2]        [,3]       [,4]
#> [1,] -0.5339738 2.901525 -0.31466300 -0.2542238
#> [2,] -1.8861745 1.646391  0.88667321 -0.3905338
#> [3,]  0.9186361 5.087257 -1.96586601  0.2533629
#> [4,] -0.4603238 1.114381  0.41905625 -1.0470477
#> [5,] -1.5735951 3.463157 -0.04853725  0.2993833
#> 
#> $layer_1$output
#>             [,1]       [,2]        [,3]       [,4]
#> [1,] -0.53158506   -72.6299 -0.30587352 -0.2476813
#> [2,]  0.21054968    31.6654  0.66787464 -0.3805885
#> [3,] -2.61504595 -2286.7737  0.37037773  0.3814392
#> [4,] -1.43140592   163.7199  0.43262353 -0.7588698
#> [5,] -0.06102861  -271.0512 -0.06460641  0.3160078
#> 
#> 
#> $layer_2
#> $layer_2$input
#>             [,1]      [,2]       [,3]        [,4]
#> [1,]   -58.58033  27.96176   40.16512   -34.66948
#> [2,]    28.36224 -13.49444  -16.50852    14.33548
#> [3,] -1887.65876 870.96099 1273.72386 -1136.53410
#> [4,]   136.50138 -63.74116  -92.96963    83.08216
#> [5,]  -222.14986 103.59979  151.65180  -134.85484
#> 
#> $layer_2$output
#>              [,1]         [,2]        [,3]       [,4]
#> [1,]   -188526631    -495025.4    -9194290   -2670784
#> [2,]   1895761173    4925743.5    91079578   27196885
#> [3,] -44911813022 -115584161.7 -2155216415 -647790945
#> [4,]   8092954815   20040456.2   380656400  121560619
#> [5,]  -2718157487   -6720729.4  -128236249  -40843813
#> 
#> 
#> $layer_3
#> $layer_3$input
#>            [,1]        [,2]         [,3]         [,4]
#> [1,]   -3995903    27795378   -188917498   -139676500
#> [2,]   38297290  -279384239   1901248955   1403971229
#> [3,] -905300430  6621792382 -45048275035 -33260285927
#> [4,]  153809736 -1195921890   8128920452   5990438126
#> [5,]  -52238799   401830488  -2729911865  -2012161409
#> 
#> $layer_3$output
#>            [,1]        [,2]         [,3]         [,4]
#> [1,]   -3995903    27795378   -188917498   -139676500
#> [2,]   38297290  -279384239   1901248955   1403971229
#> [3,] -905300430  6621792382 -45048275035 -33260285927
#> [4,]  153809736 -1195921890   8128920452   5990438126
#> [5,]  -52238799   401830488  -2729911865  -2012161409
#> 
#> 

# Predict using the obtained polynomials (for chosen layers)
predict(object = final_poly, newdata = newdata, layers = c(2,3))
#> $layer_2
#> $layer_2$input
#>             [,1]      [,2]       [,3]        [,4]
#> [1,]   -58.58033  27.96176   40.16512   -34.66948
#> [2,]    28.36224 -13.49444  -16.50852    14.33548
#> [3,] -1887.65876 870.96099 1273.72386 -1136.53410
#> [4,]   136.50138 -63.74116  -92.96963    83.08216
#> [5,]  -222.14986 103.59979  151.65180  -134.85484
#> 
#> $layer_2$output
#>              [,1]         [,2]        [,3]       [,4]
#> [1,]   -188526631    -495025.4    -9194290   -2670784
#> [2,]   1895761173    4925743.5    91079578   27196885
#> [3,] -44911813022 -115584161.7 -2155216415 -647790945
#> [4,]   8092954815   20040456.2   380656400  121560619
#> [5,]  -2718157487   -6720729.4  -128236249  -40843813
#> 
#> 
#> $layer_3
#> $layer_3$input
#>            [,1]        [,2]         [,3]         [,4]
#> [1,]   -3995903    27795378   -188917498   -139676500
#> [2,]   38297290  -279384239   1901248955   1403971229
#> [3,] -905300430  6621792382 -45048275035 -33260285927
#> [4,]  153809736 -1195921890   8128920452   5990438126
#> [5,]  -52238799   401830488  -2729911865  -2012161409
#> 
#> $layer_3$output
#>            [,1]        [,2]         [,3]         [,4]
#> [1,]   -3995903    27795378   -188917498   -139676500
#> [2,]   38297290  -279384239   1901248955   1403971229
#> [3,] -905300430  6621792382 -45048275035 -33260285927
#> [4,]  153809736 -1195921890   8128920452   5990438126
#> [5,]  -52238799   401830488  -2729911865  -2012161409
#> 
#>