Predict method for nn2poly objects.

Predicted values obtained with a nn2poly object on given data.

Usage

# S3 method for class 'nn2poly'
predict(object, newdata, layers = NULL, ...)

Arguments

object

Object of class inheriting from 'nn2poly'.

newdata

Input data as matrix, vector or dataframe. Number of columns (or elements in vector) should be the number of variables in the polynomial (dimension p). Response variable to be predicted should not be included.

layers

Vector containing the chosen layers from object to be evaluated. If set to NULL, all layers are computed. Default is set to NULL.

...

Further arguments passed to or from other methods.

Value

Returns a matrix or list of matrices with the evaluation of each polynomial at each layer as given by the provided object of class nn2poly.

If object contains the polynomials of the last layer, as given by nn2poly(object, keep_layers = FALSE), then the output is a matrix with the evaluation of each data point on each polynomial. In this matrix, each column represents the evaluation of a polynomial and each column corresponds to each point in the new data to be evaluated.

If object contains all the internal polynomials also, as given by nn2poly(object, keep_layers = TRUE), then the output is a list of layers (represented by layer_i), where each one is another list with input and output elements, where each one contains a matrix with the evaluation of the "input" or "output" polynomial at the given layer, as explained in the case without internal polynomials.

Details

Internally uses eval_poly() to obtain the predictions. However, this only works with a objects of class nn2poly while eval_poly() can be used with a manually created polynomial in list form.

When object contains all the internal polynomials also, as given by nn2poly(object, keep_layers = TRUE), it is important to note that there are two polynomial items per layer (input/output). These polynomial items will also contain several polynomials of the same structure, one per neuron in the layer, stored as matrix rows in $values. Please see the NN2Poly original paper for more details.

Note also that "linear" layers will contain the same input and output results as Taylor expansion is not used and thus the polynomials are also the same. Because of this, in the situation of evaluating multiple layers we provide the final layer with "input" and "output" even if they are the same, for consistency.

See also

Examples

# Build a NN structure with random weights, with 2 (+ bias) inputs,
# 4 (+bias) neurons in the first hidden layer with "tanh" activation
# function, 4 (+bias) neurons in the second hidden layer with "softplus",
# and 1 "linear" output unit

weights_layer_1 <- matrix(rnorm(12), nrow = 3, ncol = 4)
weights_layer_2 <- matrix(rnorm(20), nrow = 5, ncol = 4)
weights_layer_3 <- matrix(rnorm(5), nrow = 5, ncol = 1)

# Set it as a list with activation functions as names
nn_object = list("tanh" = weights_layer_1,
                 "softplus" = weights_layer_2,
                 "linear" = weights_layer_3)

# Obtain the polynomial representation (order = 3) of that neural network
final_poly <- nn2poly(nn_object, max_order = 3)

# Define some new data, it can be vector, matrix or dataframe
newdata <- matrix(rnorm(10), ncol = 2, nrow = 5)

# Predict using the obtained polynomial
predict(object = final_poly, newdata = newdata)
#> [1] -21.344227  -1.110583   4.760369  17.520142   3.465613

# Change the last layer to have 3 outputs (as in a multiclass classification)
# problem
weights_layer_4 <- matrix(rnorm(20), nrow = 5, ncol = 4)

# Set it as a list with activation functions as names
nn_object = list("tanh" = weights_layer_1,
                 "softplus" = weights_layer_2,
                 "linear" = weights_layer_4)

# Obtain the polynomial representation of that neural network
# Polynomial representation of each hidden neuron is given by
final_poly <- nn2poly(nn_object, max_order = 3, keep_layers = TRUE)

# Define some new data, it can be vector, matrix or dataframe
newdata <- matrix(rnorm(10), ncol = 2, nrow = 5)

# Predict using the obtained polynomials (for all layers)
predict(object = final_poly, newdata = newdata)
#> $layer_1
#> $layer_1$input
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] -2.2982951  6.8322228  1.8701077 -6.3762748
#> [2,] -0.3059160 -2.3812127 -0.6939258  0.6545819
#> [3,] -1.2929450  0.1069224 -0.7602996 -1.1709467
#> [4,] -0.8006397  1.8367722  1.1853521 -2.6322438
#> [5,] -2.2192979 -0.8395190 -2.9351919 -0.2643359
#> 
#> $layer_1$output
#>            [,1]        [,2]       [,3]       [,4]
#> [1,]  1.4387492 -23.4022635 -4.1317274 137.490301
#> [2,] -0.2997964   1.3914959 -0.5290092  -5.177951
#> [3,] -0.7009298   0.5874071 -0.6269358  -1.741965
#> [4,] -0.6390632   1.4971654  0.6800348   4.604232
#> [5,]  0.7243468  11.9023351  1.4406594 -26.317729
#> 
#> 
#> $layer_2
#> $layer_2$input
#>            [,1]        [,2]      [,3]       [,4]
#> [1,] -72.727031 110.2325201 13.617263 -17.218756
#> [2,]   5.065385  -2.9620557  1.477528  -1.842932
#> [3,]   2.989167  -0.5401141  2.421410  -1.633603
#> [4,]  -1.901821   4.0391400  4.420749  -4.052163
#> [5,]  16.343112 -19.1144520  3.197571  -7.799348
#> 
#> $layer_2$output
#>             [,1]       [,2]      [,3]       [,4]
#> [1,] -52.1173831 46.8111395 46.594401 21.1520303
#> [2,]   4.2759609 -0.5658585  1.216993 -0.3421481
#> [3,]   2.9900519  0.3873725  2.675544  0.1314888
#> [4,]  -0.8047415  2.5322573  5.052194  0.2698241
#> [5,]  15.0024067 -9.3772319  5.580790 -2.0441074
#> 
#> 
#> $layer_3
#> $layer_3$input
#>          [,1]        [,2]      [,3]        [,4]
#> [1,] 31.41444 -126.616883 91.769585  77.6168398
#> [2,] 11.92136    5.368994  3.080517  -1.9868663
#> [3,] 13.45481    2.070456  5.517163  -0.3701399
#> [4,] 11.80241   -5.082487  8.698053   3.9712499
#> [5,] 44.04145   18.314612  4.046083 -18.4261342
#> 
#> $layer_3$output
#>          [,1]        [,2]      [,3]        [,4]
#> [1,] 31.41444 -126.616883 91.769585  77.6168398
#> [2,] 11.92136    5.368994  3.080517  -1.9868663
#> [3,] 13.45481    2.070456  5.517163  -0.3701399
#> [4,] 11.80241   -5.082487  8.698053   3.9712499
#> [5,] 44.04145   18.314612  4.046083 -18.4261342
#> 
#> 

# Predict using the obtained polynomials (for chosen layers)
predict(object = final_poly, newdata = newdata, layers = c(2,3))
#> $layer_2
#> $layer_2$input
#>            [,1]        [,2]      [,3]       [,4]
#> [1,] -72.727031 110.2325201 13.617263 -17.218756
#> [2,]   5.065385  -2.9620557  1.477528  -1.842932
#> [3,]   2.989167  -0.5401141  2.421410  -1.633603
#> [4,]  -1.901821   4.0391400  4.420749  -4.052163
#> [5,]  16.343112 -19.1144520  3.197571  -7.799348
#> 
#> $layer_2$output
#>             [,1]       [,2]      [,3]       [,4]
#> [1,] -52.1173831 46.8111395 46.594401 21.1520303
#> [2,]   4.2759609 -0.5658585  1.216993 -0.3421481
#> [3,]   2.9900519  0.3873725  2.675544  0.1314888
#> [4,]  -0.8047415  2.5322573  5.052194  0.2698241
#> [5,]  15.0024067 -9.3772319  5.580790 -2.0441074
#> 
#> 
#> $layer_3
#> $layer_3$input
#>          [,1]        [,2]      [,3]        [,4]
#> [1,] 31.41444 -126.616883 91.769585  77.6168398
#> [2,] 11.92136    5.368994  3.080517  -1.9868663
#> [3,] 13.45481    2.070456  5.517163  -0.3701399
#> [4,] 11.80241   -5.082487  8.698053   3.9712499
#> [5,] 44.04145   18.314612  4.046083 -18.4261342
#> 
#> $layer_3$output
#>          [,1]        [,2]      [,3]        [,4]
#> [1,] 31.41444 -126.616883 91.769585  77.6168398
#> [2,] 11.92136    5.368994  3.080517  -1.9868663
#> [3,] 13.45481    2.070456  5.517163  -0.3701399
#> [4,] 11.80241   -5.082487  8.698053   3.9712499
#> [5,] 44.04145   18.314612  4.046083 -18.4261342
#> 
#>